Infer: Using state monad in `instantiate`
Rather than manually plumbing the substitution map through the
recursive function calls, and avoiding using an STRef for the state.

Using the MTL, we may add a StateT layer on top of our existing
monad stack, and then immediately run it. To facilitate this, all
the operations have been generalised to rely only upon a constraint,
rather than the true stack.