Infer: Using State monad in `abstractTy`
Same pattern as in instantiate: `abstractTy` needs to keep track of
the variables it has already converted into type references. Rather
than using an STRef for the map, here I use a StateT wrapper around
the monad in use, and evaluate it with an empty map at the top level
(of the function) to get back to our desired monad.