Lexer: Introducing ByteStrings for stream input.
They are more efficient (and compact) than String types. And we can use
them to print errors.

Pursuant to these changes, Sandbox has been axed: A bug in Alex's
implementation that uses ByteString inputs means we must now build
Alex from source (included as a submodule). This then caused issues when
interacting with the sandbox, which had a knock on effect on my will to
live.