Error: BUGFIX: Missing context information
Awkward edge case: Many expressions are annotated with a parent
context (e.g. "if expression) and then several child
contexts (e.g. "then branch"). Then when an error occurs in the child
branch, the two contexts are merged to produce a message of the form "In
the then branch of the if expression").

However, if the "if expression" occurs at the top level, and the two
branches are of different types, or the condition is not a boolean,
(causing an error) then during pretty printing, the "if expression"
label is discarded in favour of the "expression" or "definition" top
level label, because they annotate the same location. To avoid this,
we only discard information pertaining to the same location if it is not
in the top level of the error context chain.