Infer: Top level error contexts
Some cyclicity checks and unifications occur out of the scope of any
location context. To fix that, we take hold of the top level location
context, and do our operations in there, and treat it as an error for
the top level expression to not be such a location context.