Type: Removing "Broken" type.
It is not valid for a broken type to be a smaller part of a larger,
compound type (i.e. a reference, or an arrow type). To prevent that from
happening at the type level, the global definition map now only
optionally holds a type: If it is `Nothing` that signals there was an
error in the checking of that statement.