[errors] Removing Unification Error Printing Bias
When printing a unification error, if it occurs nested within another,
both examples are shown, and the first unificand is labelled "Expected"
and the second "Actual" in both cases. However, blame flips when moving
from positive to negative positions, so this is not always true. To
combat this, the biased labellings have been removed when printing the
outermost unificands.